qcom: ssr: Add no_auth and pdev info to subsys notification data
Add sending the platform device pointer and the no_auth boolean
in the notif_data struct. It is left upto the drivers who have
registered for notifications about the individual subsystems to
use this information according to their need.
This change is required to enable drivers to authenticate and
bring subsystems out of reset instead of using the PIL ops.

Change-Id: I179e13768b1ede720ede542e4ca0d7376625ed0b
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>
Signed-off-by: erikcas <erikcas1972@gmail.com>