Update calls to irq_is_per_cpu
We initially picked the irq_is_per_cpu API from a patch which had not
yet been accepted upstream. In the course of the patch being upstreamed,
the name of the API was changed from irq_is_per_cpu to irq_is_percpu.
Updating these calls to the same as what is on the remote will facilitate
code maintenance.

Change-Id: I1cac25b68ff20f219e0c01e5f6f6f006f8d269e3
Signed-off-by: Ian Maund <imaund@codeaurora.org>
Signed-off-by: erikcas <erikcas1972@gmail.com>