soc: qcom: Introduce cpu power control library
Introduce a library for power controlling the cpu and l2cache
subsystem, which exposes apis utilized by arm/arm64 smp setup
code.

As a part of this patch also add a framework for powering on
the L2 caches for a particular subsystem and implement L2 cache
power-on sequences for MSM8916/39 and MSM8994 cpu subsystems.

Change-Id: I48f503cbcbef488d990f66096e8f62ed7e034742
Signed-off-by: Abhimanyu Kapur <abhimany@codeaurora.org>
Signed-off-by: erikcas <erikcas1972@gmail.com>