qcom: subsys-pil-tz: Replace using clk_get_sys for scm clocks
The usage of clk_get_sys is not mandatory for the scm clocks
any more because their usage has moved to the generic PIL
driver from the scm-pas driver.
Unlike the scm-pas driver, the subsys-pil-tz driver can use
the device node pointer of the subsystems to call clk_get on
these clocks instead.

Change-Id: I560df07293cbfb53a293c543a4cf70a2358f0310
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>
Signed-off-by: erikcas <erikcas1972@gmail.com>