ARM: dts: msm: enable UFS regulators at boot-up for MSM8994
The default power on sequence enables LDO 20, LDO 28 and LDO 31
regulators which are required to boot from UFS.

If the board is configured to use only one of the storage devices, then
the regulators corresponding to the unused device are unnecessary left
always on.

Update the configuration for the above LDOs so that the kernel enables
these regulators while it's booting up and so that it disables them
during the late_init call stage if the UFS driver has not requested
that they be left enabled.

Change-Id: I46e392eb1a3a91cb1d28bc4f1b3dbf602a0c0359
Signed-off-by: David Keitel <dkeitel@codeaurora.org>
Signed-off-by: erikcas <erikcas1972@gmail.com>