msm: kgsl: Map full GPU address range and access shader memory as an offset
On older GPU's GPU SMMU registers were physically located between GPU
registers and GPU shader memory, so we had to map these 2 physical regions
seperately in driver. On newer GPU's this is not a limitation anymore
so map full GPU address range and access shader memory as an offset
when required.

Change-Id: I3b17595e631a1f4fa2d913214ebc3911f4d2e3d9
Signed-off-by: Tarun Karra <tkarra@codeaurora.org>
Signed-off-by: erikcas <erikcas1972@gmail.com>