regulator: cpr-regulator: remove unnecessary pointer checks in callbacks
Remove cpr_vreg pointer NULL checks in cpr-regulator debugfs
callback functions.  The cpr_vreg pointer is guaranteed to be
valid in these callback functions so there is no need to check
it.  Additionally, the print statements inside of the NULL checks
where incorrectly referencing cpr_vreg after it was known to be
NULL.

Change-Id: Id2efd9f28f97ce448fd86f03922ca2cfd63173a8
Signed-off-by: David Collins <collinsd@codeaurora.org>
Signed-off-by: erikcas <erikcas1972@gmail.com>